pyproject.toml
[project]
name = "hse-applied-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "geopy>=2.4.1",
    "openmeteo-requests>=1.7.4",
    "python-dotenv>=1.2.1",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "seaborn>=0.13.2",
    "streamlit>=1.52.1",
    "timezonefinder>=8.1.0",
]
